+++ .eslintrc.js
@@ -6,7 +6,6 @@ module.exports = {
     ],
     parser: "@typescript-eslint/parser",
     parserOptions: {
-        project: 'tsconfig.json',
         sourceType: "module",
         ecmaVersion: 2018
     },
