+++ package.json
@@ -11,7 +11,7 @@
     "type-check": "tsc --noEmitOnError",
     "type-check:watch": "npm run type-check -- --watch",
     "build": "npm run build:js",
-    "build:js": "babel src --out-dir dist --extensions \".ts\" --source-maps inline && tsc --watch"
+    "build:js": "babel src --out-dir dist && babel src/bin/www --copy-files --out-dir dist/bin"
   },
   "repository": {
     "type": "git",
