+++ src/routes/add-ons/autocomplete.js
@@ -1,7 +1,7 @@
 import express from 'express';
 import asyncHandler from 'express-async-handler';
 import { getResources } from '@controllers/add-ons/autocomplete/resource';
-import { getDistinctValues } from '@controllers/add-ons/autocomplete/distinct';
+import { getDistinctValues } from '@/controllers/add-ons/autocomplete/distinct';
 
 const router = express.Router();
